Hw 14 (#58)
* add basic auth guard

* add blog body validation

* add posts body validation

* add users body validation

* add HttpExceptionFilter & GlobalValidationPipe

* small refactor HttpExceptionFilter

* add rate limit to auth & create folder for local/jwt strategies

* add custom trim decorator

* add auth dto

* add email adapter & email manager

* change login return status

* set config

* set esLint config

* add registration route & test

* add registration-confirmation route & tests & delete passport strategies & refactor user schema

* refactor updateConfirmEmailStatus in users repo

* add resending email route & tests

* add login fn & tests

* add cookie parser

* fix login endpoint

* fix login endpoint #2

* fix login endpoint #2

* add fix

* fix confirm register endpoint

* fix

* fix users repo

* fix users pagination & user schema

* fix users pagination in query repo & user schema

* fix users

* add blog/post exist decorators

* fix blog exist decorator & add blog endpoint tests

* add post like fn

* refactor post like status fn & create/get comment by post

* add comments update, delete, update like status fn & tests

* fix blogId exist decorator

* fix create post by blog decorator validation

* add check if comment not found

* add try cathc for get user by id

* fix double useGuards

* refactor comments double guards

* fix comment like status endpoint if comment not found

* refactor update like method in comment repo

* add 403 exception in global filter

* add check if the user is not comfirmed then he should not login

* add get me endpoint

* add devices fn in login endpoint

* add refresh token endpoint fn

* add logout endpoint fn

* add password recovery enpoint fn

* add new password endpoint fn

* add get all user devices endpoint fn

* add delete devices endpoints fn

* fix device lastActiveDate in refresh token endpoint

* refactor checkDevice

* add check lastActiveDate in refresh token guard

* remove check lastActiveDate in refresh token guard

* fix check device

+++ src/modules/auth/auth.service.ts
@@ -6,6 +6,7 @@ import { Injectable } from '@nestjs/common';
 import { addMinutesToCurrentDate } from '../../helpers/add.minutes.current.date';
 import { ReqUserType } from '../../types/req.user.interface';
 import { DevicesService } from '../devices/devices.service';
+import { DevicesRepo } from '../devices/repositories/devices.repo';
 import { EmailService } from '../email/email.service';
 import { UsersRepo } from '../users/repositories/users.repo';
 import { UsersService } from '../users/users.service';
@@ -25,6 +26,7 @@ export class AuthService {
     private usersRepo: UsersRepo,
     private jwtService: JwtService,
     private devicesService: DevicesService,
+    private devicesRepo: DevicesRepo,
   ) {}
 
   async register(dto: RegisterDto): Promise<void> {
@@ -85,7 +87,13 @@ export class AuthService {
   }
 
   async refreshToken(refreshPayload: ReqUserType, ip: string) {
-    const { userId, deviceId } = refreshPayload;
+    const { userId, deviceId, iat } = refreshPayload;
+
+    const device = await this.devicesRepo.getDeviceById(deviceId);
+    if (!device) return null;
+    if (device.userId !== userId) return null;
+    if (device.lastActiveDate !== new Date(iat * 1000).toISOString())
+      return null;
 
     const tokens = this.jwtService.createJWT(userId, deviceId);
 
@@ -102,6 +110,11 @@ export class AuthService {
   }
 
   async logout(dto: ReqUserType) {
+    const device = await this.devicesRepo.getDeviceById(dto.deviceId);
+    if (!device) return null;
+    if (device.userId !== dto.userId) return null;
+    if (device.lastActiveDate !== new Date(dto.iat * 1000).toISOString())
+      return null;
     return await this.devicesService.deleteOneDevice(dto);
   }
