+++ src/modules/devices/devices.controller.ts
@@ -7,6 +7,7 @@ import {
   HttpStatus,
   NotFoundException,
   Param,
+  UnauthorizedException,
   UseGuards,
 } from '@nestjs/common';
 
@@ -15,6 +16,7 @@ import { RefreshTokenGuard } from '../../infra/guards/auth/refresh.token.guard';
 import { ReqUserType } from '../../types/req.user.interface';
 import { DevicesService } from './devices.service';
 import { DevicesQueryRepo } from './repositories/devices.query.repo';
+import { DevicesRepo } from './repositories/devices.repo';
 
 @Controller('security')
 @UseGuards(RefreshTokenGuard)
@@ -22,12 +24,18 @@ export class DevicesController {
   constructor(
     private devicesQueryRepo: DevicesQueryRepo,
     private devicesService: DevicesService,
+    private devicesRepo: DevicesRepo,
   ) {}
 
   @Get('/devices')
   async getAllUserDevices(@RefreshTokenPayload() refreshPayload: ReqUserType) {
-    const { userId } = refreshPayload;
+    const { userId, deviceId, iat } = refreshPayload;
 
+    const device = await this.devicesRepo.getDeviceById(deviceId);
+    if (!device) throw new UnauthorizedException();
+    if (device.userId !== userId) throw new UnauthorizedException();
+    if (device.lastActiveDate !== new Date(iat * 1000).toISOString())
+      throw new UnauthorizedException();
     return await this.devicesQueryRepo.getAllUserDevices(userId);
   }
 
@@ -36,6 +44,13 @@ export class DevicesController {
   async deleteAllDevicesExceptCurrent(
     @RefreshTokenPayload() refreshPayload: ReqUserType,
   ) {
+    const { userId, deviceId, iat } = refreshPayload;
+
+    const device = await this.devicesRepo.getDeviceById(deviceId);
+    if (!device) throw new UnauthorizedException();
+    if (device.userId !== userId) throw new UnauthorizedException();
+    if (device.lastActiveDate !== new Date(iat * 1000).toISOString())
+      throw new UnauthorizedException();
     return await this.devicesService.deleteAllDevicesExceptCurrent(
       refreshPayload,
     );
@@ -43,17 +58,17 @@ export class DevicesController {
 
   @Delete('/devices/:deviceId')
   @HttpCode(HttpStatus.NO_CONTENT)
-  async DeleteOneDeviceDto(
+  async DeleteOneDevice(
     @Param('deviceId') deviceId: string,
     @RefreshTokenPayload() refreshPayload: ReqUserType,
   ) {
     const { userId } = refreshPayload;
 
-    const device = await this.devicesQueryRepo.getDevicebyId(deviceId);
-
+    const device = await this.devicesRepo.getDeviceById(deviceId);
     if (!device) throw new NotFoundException();
-
     if (device.userId !== userId) throw new ForbiddenException();
+    // if (device.lastActiveDate !== new Date(iat * 1000).toISOString())
+    //   throw new UnauthorizedException();
 
     return await this.devicesService.deleteOneDevice({ deviceId, userId });
   }
